+++ backend/src/main/java/gov/cdc/usds/simplereport/api/patient/PatientExportController.java
@@ -1,9 +1,7 @@
 package gov.cdc.usds.simplereport.api.patient;
 
 import gov.cdc.usds.simplereport.config.AuthorizationConfiguration;
-import gov.cdc.usds.simplereport.config.AuthorizationConfiguration.RequirePermissionEditPatientAtFacility;
-import gov.cdc.usds.simplereport.service.FacilityCsvExportService;
-import gov.cdc.usds.simplereport.service.TestOrderService;
+import gov.cdc.usds.simplereport.service.CsvExportService;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.UUID;
@@ -20,31 +18,24 @@ import org.springframework.web.servlet.mvc.method.annotation.StreamingResponseBo
 @Slf4j
 @RequiredArgsConstructor
 public class PatientExportController {
-  private final FacilityCsvExportService csvExportService;
+  private final CsvExportService csvExportService;
 
-  @RequirePermissionEditPatientAtFacility
+  @AuthorizationConfiguration.RequirePermissionCreatePatientAtFacility
   @GetMapping(value = "/patients/download/facility")
   public ResponseEntity<StreamingResponseBody> downloadFacilityPatientsAsCSV(
-      @RequestParam() UUID facilityId,
-      @RequestParam(defaultValue = "0") int pageNumber,
-      @RequestParam(defaultValue = "5000") int pageSize) {
+      @RequestParam() UUID facilityId) {
 
     log.info("Facility patients CSV download request for facilityId={}", facilityId);
 
-    if (pageNumber < 0) {
-      pageNumber = TestOrderService.DEFAULT_PAGINATION_PAGEOFFSET;
-    }
-    if (pageSize < 1) {
-      pageSize = TestOrderService.DEFAULT_PAGINATION_PAGESIZE;
-    }
+    String timestamp = new SimpleDateFormat("yyyyMMdd-HHmmss").format(new Date());
+    String zippedCsvFileName = String.format("facility-patients-%s.zip", timestamp);
+    String unZippedCsvFileName = String.format("facility-patients-%s.csv", timestamp);
 
     StreamingResponseBody responseBody =
         outputStream -> {
-          csvExportService.streamFacilityPatientsAsZippedCsv(outputStream, facilityId);
+          csvExportService.streamFacilityPatientsAsZippedCsv(
+              outputStream, facilityId, unZippedCsvFileName);
         };
-
-    String timestamp = new SimpleDateFormat("yyyyMMdd-HHmmss").format(new Date());
-    String zippedCsvFileName = String.format("facility-%s-patients-%s.zip", facilityId, timestamp);
 
     return ResponseEntity.ok()
         .header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=" + zippedCsvFileName)
@@ -55,26 +46,19 @@ public class PatientExportController {
   @AuthorizationConfiguration.RequirePermissionToAccessOrg
   @GetMapping(value = "/patients/download/organization")
   public ResponseEntity<StreamingResponseBody> downloadOrganizationPatientsAsCSV(
-      @RequestParam() UUID orgId,
-      @RequestParam(defaultValue = "0") int pageNumber,
-      @RequestParam(defaultValue = "5000") int pageSize) {
+      @RequestParam() UUID orgId) {
 
     log.info("Organization patients CSV download request for organizationId={}", orgId);
 
-    if (pageNumber < 0) {
-      pageNumber = TestOrderService.DEFAULT_PAGINATION_PAGEOFFSET;
-    }
-    if (pageSize < 1) {
-      pageSize = TestOrderService.DEFAULT_PAGINATION_PAGESIZE;
-    }
+    String timestamp = new SimpleDateFormat("yyyyMMdd-HHmmss").format(new Date());
+    String zippedCsvFileName = String.format("organization-patients-%s.zip", timestamp);
+    String unZippedCsvFileName = String.format("organization-patients-%s.csv", timestamp);
 
     StreamingResponseBody responseBody =
         outputStream -> {
-          csvExportService.streamOrganizationPatientsAsZippedCsv(outputStream, orgId);
+          csvExportService.streamOrganizationPatientsAsZippedCsv(
+              outputStream, orgId, unZippedCsvFileName);
         };
-
-    String timestamp = new SimpleDateFormat("yyyyMMdd-HHmmss").format(new Date());
-    String zippedCsvFileName = String.format("organization-%s-patients-%s.zip", orgId, timestamp);
 
     return ResponseEntity.ok()
         .header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=" + zippedCsvFileName)
